Cloudfunction/requirements.txt
functions-framework==3.*
joblib==1.5.2
flask==2.3.2
pandas==2.3.2
xgboost==3.0.5
google-cloud-storage==2.11.0
scikit-learn==1.5.2